IGNITE-24792 Move some classes to commons
