+++ modules/commons/src/main/java/org/apache/ignite/internal/util/GridCommonFunc.java
@@ -1,0 +1,174 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.ignite.internal.util;
+
+import java.util.Collection;
+import java.util.Map;
+import org.apache.ignite.internal.util.lang.gridfunc.StringConcatReducer;
+import org.apache.ignite.internal.util.typedef.internal.A;
+import org.apache.ignite.lang.IgniteReducer;
+import org.jetbrains.annotations.Nullable;
+
+/**
+ * Contains factory and utility methods for {@code closures}, {@code predicates}, and {@code tuples}.
+ * It also contains functional style collection comprehensions.
+ * <p>
+ * Most of the methods in this class can be divided into two groups:
+ * <ul>
+ * <li><b>Factory</b> higher-order methods for closures, predicates and tuples, and</li>
+ * <li><b>Utility</b> higher-order methods for processing collections with closures and predicates.</li>
+ * </ul>
+ * Note that contrary to the usual design this class has substantial number of
+ * methods (over 200). This design is chosen to simulate a global namespace
+ * (like a {@code Predef} in Scala) to provide general utility and functional
+ * programming functionality in a shortest syntactical context using {@code F}
+ * typedef.
+ * <p>
+ * Also note, that in all methods with predicates, null predicate has a {@code true} meaning. So does
+ * the empty predicate array.
+ *
+ * <b>Remove when GridFunc migrated</b>
+ */
+public class GridCommonFunc {
+    /**
+     * Concatenates strings using provided delimiter.
+     *
+     * @param c Input collection.
+     * @param delim Delimiter (optional).
+     * @return Concatenated string.
+     */
+    public static String concat(Iterable<?> c, @Nullable String delim) {
+        A.notNull(c, "c");
+
+        IgniteReducer<? super String, String> f = new StringConcatReducer(delim);
+
+        for (Object x : c)
+            if (!f.collect(x == null ? null : x.toString()))
+                break;
+
+        return f.reduce();
+    }
+
+    /**
+     * Tests if given string is {@code null} or empty.
+     *
+     * @param s String to test.
+     * @return Whether or not the given string is {@code null} or empty.
+     */
+    public static boolean isEmpty(@Nullable String s) {
+        return s == null || s.isEmpty();
+    }
+
+    /**
+     * Tests if the given array is either {@code null} or empty.
+     *
+     * @param c Array to test.
+     * @return Whether or not the given array is {@code null} or empty.
+     */
+    public static <T> boolean isEmpty(@Nullable T[] c) {
+        return c == null || c.length == 0;
+    }
+
+    /**
+     * Tests if the given array is {@code null}, empty or contains only {@code null} values.
+     *
+     * @param c Array to test.
+     * @return Whether or not the given array is {@code null}, empty or contains only {@code null} values.
+     */
+    public static <T> boolean isEmptyOrNulls(@Nullable T[] c) {
+        if (isEmpty(c))
+            return true;
+
+        for (T element : c)
+            if (element != null)
+                return false;
+
+        return true;
+    }
+
+    /**
+     * Tests if the given array is either {@code null} or empty.
+     *
+     * @param c Array to test.
+     * @return Whether or not the given array is {@code null} or empty.
+     */
+    public static boolean isEmpty(@Nullable int[] c) {
+        return c == null || c.length == 0;
+    }
+
+    /**
+     * Tests if the given array is either {@code null} or empty.
+     *
+     * @param c Array to test.
+     * @return Whether or not the given array is {@code null} or empty.
+     */
+    public static boolean isEmpty(@Nullable byte[] c) {
+        return c == null || c.length == 0;
+    }
+
+    /**
+     * Tests if the given array is either {@code null} or empty.
+     *
+     * @param c Array to test.
+     * @return Whether or not the given array is {@code null} or empty.
+     */
+    public static boolean isEmpty(@Nullable long[] c) {
+        return c == null || c.length == 0;
+    }
+
+    /**
+     * Tests if the given array is either {@code null} or empty.
+     *
+     * @param c Array to test.
+     * @return Whether or not the given array is {@code null} or empty.
+     */
+    public static boolean isEmpty(@Nullable char[] c) {
+        return c == null || c.length == 0;
+    }
+
+    /**
+     * Tests if the given collection is either {@code null} or empty.
+     *
+     * @param c Collection to test.
+     * @return Whether or not the given collection is {@code null} or empty.
+     */
+    public static boolean isEmpty(@Nullable Iterable<?> c) {
+        return c == null || (c instanceof Collection<?> ? ((Collection<?>)c).isEmpty() : !c.iterator().hasNext());
+    }
+
+    /**
+     * Tests if the given collection is either {@code null} or empty.
+     *
+     * @param c Collection to test.
+     * @return Whether or not the given collection is {@code null} or empty.
+     */
+    public static boolean isEmpty(@Nullable Collection<?> c) {
+        return c == null || c.isEmpty();
+    }
+
+    /**
+     * Tests if the given map is either {@code null} or empty.
+     *
+     * @param m Map to test.
+     * @return Whether or not the given collection is {@code null} or empty.
+     */
+    public static boolean isEmpty(@Nullable Map<?, ?> m) {
+        return m == null || m.isEmpty();
+    }
+
+}
